[0.7.1-test] Small Tidy.

+++ src/io/siri/joe/Maths.java
@@ -118,6 +118,6 @@ public class Maths {
             nn = nn >> 1;
             i++;
         }
-        return (double)i;
+        return i;
     }
 }
